requirements.txt
python-telegram-bot==20.3
fastapi==0.94.1
uvicorn==0.22.0
sqlalchemy==2.0.14
openai==0.27.0
pillow==9.5.0 